+++ deployables/app/app/routes/edit/list-routes.tsx
@@ -1,8 +1,7 @@
-import { fromVersion, MinimumVersion, OnlyConnected, Page } from '@/components'
-import { useIsPending } from '@/hooks'
-import { Chain } from '@/routes-ui'
+import { fromVersion, OnlyConnected, Page } from '@/components'
+import { dbClient, getAccounts, type Account } from '@/db'
 import { routeTitle } from '@/utils'
-import { CHAIN_NAME, getChainId, ZERO_ADDRESS } from '@zodiac/chains'
+import { authKitLoader } from '@/workOS/server'
 import { getString } from '@zodiac/form-data'
 import {
   CompanionAppMessageType,
@@ -10,45 +9,61 @@ import {
   CompanionResponseMessageType,
   useExtensionMessageHandler,
 } from '@zodiac/messages'
-import { type ExecutionRoute } from '@zodiac/schema'
 import {
-  Address,
-  Form,
-  GhostButton,
-  GhostLinkButton,
   Info,
-  MeatballMenu,
-  Modal,
-  PrimaryButton,
   SecondaryLinkButton,
   Table,
   TableBody,
-  TableCell,
   TableHead,
   TableHeader,
   TableRow,
-  Tag,
 } from '@zodiac/ui'
-import classNames from 'classnames'
-import { Pencil, Play, Trash2 } from 'lucide-react'
-import { useEffect, useState, type PropsWithChildren } from 'react'
-import { href, useRevalidator } from 'react-router'
+import { type PropsWithChildren } from 'react'
+import { useRevalidator } from 'react-router'
 import type { Route } from './+types/list-routes'
 import { Intent } from './intents'
 import { loadActiveRouteId } from './loadActiveRouteId'
 import { loadRoutes } from './loadRoutes'
+import { LocalAccount } from './LocalAccount'
+import { RemoteAccount } from './RemoteAccount'
 
 export const meta: Route.MetaFunction = ({ matches }) => [
   { title: routeTitle(matches, 'Accounts') },
 ]
 
-export const clientLoader = async () => {
-  const [routes, activeRouteId] = await Promise.all([
+export const loader = (args: Route.LoaderArgs) =>
+  authKitLoader(
+    args,
+    async ({
+      context: {
+        auth: { user },
+      },
+    }) => {
+      if (user == null) {
+        return {
+          remoteAccounts: [] as Account[],
+        }
+      }
+
+      return {
+        remoteAccounts: await getAccounts(dbClient(), {
+          tenantId: user.tenantId,
+          userId: user.id,
+        }),
+      }
+    },
+  )
+
+export const clientLoader = async ({
+  serverLoader,
+}: Route.ClientLoaderArgs) => {
+  const [localAccounts, { remoteAccounts }, activeRouteId] = await Promise.all([
     loadRoutes(),
+    serverLoader(),
     fromVersion('3.6.0', () => loadActiveRouteId()),
   ])
 
-  return { routes, activeRouteId }
+  return { localAccounts, remoteAccounts, activeRouteId }
 }
 
 clientLoader.hydrate = true as const
@@ -92,13 +107,19 @@ export const clientAction = async ({ request }: Route.ClientActionArgs) => {
   }
 }
 
-const ListRoutes = ({ loaderData }: Route.ComponentProps) => {
+const ListRoutes = ({
+  loaderData: { remoteAccounts, ...clientData },
+}: Route.ComponentProps) => {
   const { revalidate, state } = useRevalidator()
 
   useExtensionMessageHandler(
     CompanionResponseMessageType.PROVIDE_ACTIVE_ROUTE,
     ({ activeRouteId }) => {
-      if (state === 'idle' && activeRouteId !== loaderData.activeRouteId) {
+      if (
+        state === 'idle' &&
+        'activeRouteId' in clientData &&
+        activeRouteId !== clientData.activeRouteId
+      ) {
         revalidate()
       }
     },
@@ -109,18 +130,26 @@ const ListRoutes = ({ loaderData }: Route.ComponentProps) => {
       <Page.Header>Accounts</Page.Header>
 
       <Page.Main>
+        {remoteAccounts.length > 0 && (
+          <Accounts>
+            {remoteAccounts.map((account) => (
+              <RemoteAccount key={account.id} account={account} />
+            ))}
+          </Accounts>
+        )}
+
         <OnlyConnected>
-          {'routes' in loaderData &&
-            (loaderData.routes.length > 0 ? (
-              <Routes>
-                {loaderData.routes.map((route) => (
-                  <Route
+          {'localAccounts' in clientData &&
+            (clientData.localAccounts.length > 0 ? (
+              <Accounts>
+                {clientData.localAccounts.map((route) => (
+                  <LocalAccount
                     key={route.id}
                     route={route}
-                    active={route.id === loaderData.activeRouteId}
+                    active={route.id === clientData.activeRouteId}
                   />
                 ))}
-              </Routes>
+              </Accounts>
             ) : (
               <Info title="You haven't created any accounts, yet.">
                 Accounts let you quickly impersonate other safes and record
@@ -140,7 +169,7 @@ const ListRoutes = ({ loaderData }: Route.ComponentProps) => {
 
 export default ListRoutes
 
-const Routes = ({ children }: PropsWithChildren) => {
+const Accounts = ({ children }: PropsWithChildren) => {
   return (
     <Table
       bleed
@@ -163,170 +192,5 @@ const Routes = ({ children }: PropsWithChildren) => {
 
       <TableBody>{children}</TableBody>
     </Table>
-  )
-}
-
-type RouteProps = { route: ExecutionRoute; active: boolean }
-
-const Route = ({ route, active }: RouteProps) => {
-  const chainId = getChainId(route.avatar)
-
-  return (
-    <TableRow
-      className="group"
-      href={href('/edit/:routeId', { routeId: route.id })}
-    >
-      <TableCell aria-describedby={route.id}>{route.label}</TableCell>
-      <TableCell>
-        {active && (
-          <Tag aria-hidden id={route.id} color="green">
-            Active
-          </Tag>
-        )}
-      </TableCell>
-      <TableCell>
-        <Chain chainId={chainId}>{CHAIN_NAME[chainId]}</Chain>
-      </TableCell>
-      <TableCell>
-        {route.initiator == null ? (
-          <Address>{ZERO_ADDRESS}</Address>
-        ) : (
-          <Address shorten>{route.initiator}</Address>
-        )}
-      </TableCell>
-      <TableCell>
-        <Address shorten>{route.avatar}</Address>
-      </TableCell>
-      <TableCell>
-        <Actions routeId={route.id} />
-      </TableCell>
-    </TableRow>
-  )
-}
-
-const Actions = ({ routeId }: { routeId: string }) => {
-  const submitting = useIsPending(
-    undefined,
-    (data) => data.get('routeId') === routeId,
-  )
-  const [menuOpen, setMenuOpen] = useState(false)
-  const [confirmingDelete, setConfirmingDelete] = useState(false)
-
-  return (
-    <div
-      className={classNames(
-        'flex justify-end gap-1 transition-opacity group-hover:opacity-100',
-        submitting || menuOpen ? 'opacity-100' : 'opacity-0',
-      )}
-    >
-      <MinimumVersion version="3.6.0">
-        <Launch routeId={routeId} />
-      </MinimumVersion>
-      <MeatballMenu
-        open={menuOpen || confirmingDelete}
-        size="tiny"
-        label="Account options"
-        onRequestShow={() => setMenuOpen(true)}
-        onRequestHide={() => setMenuOpen(false)}
-      >
-        <Edit routeId={routeId} />
-
-        <MinimumVersion version="3.6.0">
-          <Delete routeId={routeId} onConfirmChange={setConfirmingDelete} />
-        </MinimumVersion>
-      </MeatballMenu>
-    </div>
-  )
-}
-
-const Launch = ({ routeId }: { routeId: string }) => {
-  const submitting = useIsPending(
-    Intent.Launch,
-    (data) => data.get('routeId') === routeId,
-  )
-
-  return (
-    <Form intent={Intent.Launch}>
-      <GhostButton
-        submit
-        align="left"
-        size="tiny"
-        name="routeId"
-        value={routeId}
-        busy={submitting}
-        icon={Play}
-      >
-        Launch
-      </GhostButton>
-    </Form>
-  )
-}
-
-const Edit = ({ routeId }: { routeId: string }) => (
-  <GhostLinkButton
-    to={href('/edit/:routeId', { routeId })}
-    align="left"
-    size="tiny"
-    icon={Pencil}
-  >
-    Edit
-  </GhostLinkButton>
-)
-
-const Delete = ({
-  routeId,
-  onConfirmChange,
-}: {
-  routeId: string
-  onConfirmChange: (state: boolean) => void
-}) => {
-  const [confirmDelete, setConfirmDelete] = useState(false)
-  const submitting = useIsPending(
-    Intent.Delete,
-    (data) => data.get('routeId') === routeId,
-  )
-
-  useEffect(() => {
-    onConfirmChange(confirmDelete)
-  }, [confirmDelete, onConfirmChange])
-
-  return (
-    <>
-      <GhostButton
-        align="left"
-        size="tiny"
-        icon={Trash2}
-        style="critical"
-        onClick={() => setConfirmDelete(true)}
-        busy={submitting}
-      >
-        Delete
-      </GhostButton>
-
-      <Modal
-        title="Confirm delete"
-        closeLabel="Cancel"
-        onClose={() => setConfirmDelete(false)}
-        open={confirmDelete}
-        description="Are you sure you want to delete this account? This action cannot be undone."
-      >
-        <Form intent={Intent.Delete} onSubmit={() => setConfirmDelete(false)}>
-          <Modal.Actions>
-            <PrimaryButton
-              submit
-              name="routeId"
-              value={routeId}
-              busy={submitting}
-            >
-              Delete
-            </PrimaryButton>
-
-            <GhostButton onClick={() => setConfirmDelete(false)}>
-              Cancel
-            </GhostButton>
-          </Modal.Actions>
-        </Form>
-      </Modal>
-    </>
   )
 }
