+++ deployables/app/workOS/server/authKitLoader.ts
@@ -12,7 +12,7 @@ export async function authKitLoader<
 >(
   args: Args,
   fn: Fn,
-  options: GetAuthOptions,
+  options?: GetAuthOptions,
 ): Promise<Awaited<ReturnType<typeof fn>>>
 export async function authKitLoader<
   Args extends LoaderFunctionArgs,
